Re-Added Dependency on SimpleOptions
- The dependency on SimpleOptions was intentional.

SimpleOptions was intended to be the flagship package where both an interface for both environments and the options available in them was defined, and SimpleEnvs was meant to be a separate repository of environments implemented using that interface.

I might potentially revisit this in the future - I have a rough plan to define an option-free environment interface here, then add an option-enabled subclass of it in SimpleOptions - but it's not a priority for right now.